requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-core>=0.1.0
pulp>=2.7.0